% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download BDMEP files}
\usage{
download_data(year, destpath, quiet = FALSE, unzip_file = TRUE)
}
\arguments{
\item{year}{numeric. From 200 to current year.}

\item{destpath}{character. Destination path to the file.}

\item{quiet}{logical. Print messages and progress bar on console.}

\item{unzip_file}{logical. Unzip file after downloading.}
}
\value{
The file list of the destination path.
}
\description{
This function download BDMEP files from INMET
}
