% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bdmep.R
\name{parse_bdmep}
\alias{parse_bdmep}
\title{Parse BDMEP files}
\usage{
parse_bdmep(file)
}
\arguments{
\item{file}{character. File path of the weather station data.}
}
\value{
A tibble.
}
\description{
This function parse BDMEP files from INMET
}
